@@ -3,12 +3,12 @@ apply plugin: 'com.github.dcendents.android-maven'
 
 group='com.github.AlphaBoom'
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 15
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -42,6 +42,18 @@ javadoc {
     }
 }
 
+//根据Library名称生存jar包到build目录下
+//可根据自己需求更改
+task againMakeJar(type: Copy) {
+    def name = project.name //Library名称
+    delete 'libs/' + name + '.jar' //删除之前的旧jar包
+    from('build/intermediates/packaged-classes/release/') //从这个目录下取出默认jar包
+    into('libs/') //将jar包输出到指定目录下
+    include('classes.jar')
+    rename('classes.jar', name + '.jar') //自定义jar包的名字
+}
+againMakeJar.dependsOn(build)
+
 // build a jar with javadoc
 task javadocJar(type: Jar, dependsOn: javadoc) {
     classifier = 'javadoc'
@@ -55,5 +67,5 @@ artifacts {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     testImplementation 'junit:junit:4.12'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
 }
