@@ -1,6 +1,6 @@
 package com.anarchy.classifyview.sample.ireader.model;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
 import java.util.List;
