@@ -1,14 +1,13 @@
 package com.anarchy.classifyview.sample.layoutmanager;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.ViewGroup;
 
-import com.anarchy.classify.ClassifyItemAnimator;
 import com.anarchy.classify.ClassifyView;
 
 /**
