@@ -1,6 +1,6 @@
 package com.anarchy.classifyview.sample.normalfolder.logic;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
