@@ -2,7 +2,7 @@ package com.anarchy.classify.adapter;
 
 import android.app.Dialog;
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.VelocityTracker;
 import android.view.View;
 
