@@ -1,7 +1,7 @@
 package com.anarchy.classify.callback;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.VelocityTracker;
 import android.view.View;
 
