@@ -1,6 +1,6 @@
 package com.anarchy.classify.callback;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.anarchy.classify.MergeInfo;
