@@ -1,7 +1,7 @@
 package com.anarchy.classify.simple;
 
 import android.app.Dialog;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.view.ViewGroup;
 
