@@ -6,7 +6,7 @@ import android.animation.PropertyValuesHolder;
 import android.animation.ValueAnimator;
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.support.v4.widget.ScrollerCompat;
+import androidx.core.widget.ScrollerCompat;
 import android.util.AttributeSet;
 import android.util.Log;
 import android.view.View;
